+++ loveship/components/index.ts
@@ -1,7 +1,6 @@
 export * from './buttons';
 export * from './layout';
 export * from './typography';
-export * from './types';
 export * from './overlays';
 export * from './inputs';
 export * from './widgets';
@@ -15,6 +14,6 @@ export { FiltersPanel, PresetsPanel, defaultPredicates, Calendar, NumericInput, 
     SuccessNotification, HintNotification, ErrorNotification, ClearNotification, FlexSpacer, FlexCell, Spinner, DataPickerBody, PickerModal, DataPickerFooter,
     ModalBlocker, ModalHeader, ModalFooter, LabeledInput, DropdownMenuBody, DropdownMenuButton, DropdownMenuSplitter, DropdownMenuHeader,
     DropdownSubMenu, DropdownMenuSwitchButton, TimePicker, FileCard, SvgCircleProgress, DropSpot, DropMarker, IndeterminateBar, ProgressBar, IndicatorBar, Slider,
-    RichTextView, ScrollBars, VirtualList, Checkbox, TextInput, SearchInput, RadioInput,
+    RichTextView, ScrollBars, VirtualList, Checkbox, TabButton, VerticalTabButton, RadioInput, TextInput, SearchInput,
 } from '@epam/uui';
 export { MainMenuLogo, MainMenuCustomElement } from '@epam/uui-components';