+++ uui-build/scripts/ts/trackBundleSize.ts
@@ -1,0 +1,3 @@
+import { main } from '../../ts/trackBundleSize/main';
+
+main();