+++ uui-build/ts/figmaTokensGen/constants.ts
@@ -1,0 +1,6 @@
+export const IGNORED_VAR_PLACEHOLDER = '/* UUI internal variable */';
+
+export const PATH = {
+    FIGMA_VARS_COLLECTION_SRC: 'public/docs/figmaTokensGen/Theme.json',
+    RESULTS_DIR: 'public/docs/figmaTokensGen',
+};