+++ uui-build/ts/trackBundleSize/steps/buildAllModules.ts
@@ -1,0 +1,5 @@
+import { runYarnScriptFromRootSync } from '../../jsBridge';
+
+export const buildAllModules = async () => {
+    runYarnScriptFromRootSync('build-modules');
+};