+++ uui/components/widgets/Spinner.tsx
@@ -4,7 +4,7 @@ import css from './Spinner.module.scss';
 
 export type SpinnerProps = UuiSpinnerProps;
 
-export function applySpinnerMods() {
+function applySpinnerMods() {
     return [css.root, 'uui-spinner'];
 }