+++ changelog.md
@@ -1,6 +1,7 @@
-# 5.xx.xx - xx.xx.2023
+# 5.x.x - xx.xx.xxxx
 
 **What's New**
+* Added semantic values to color props 'Loveship' & 'Promo' skins
 * [NumericInput]: size `48` is deprecated and will be removed in future release. Please, use size `42` instead.
 * [Pickers]: turn on flatten search results by default for all Pickers. Now for pickers with tree structure and in search mode, results will be shown as a flat tree, each item will have a subtitle with parents path.
 * [RadioInput]: in Loveship skin removed property 'theme'.
@@ -10,6 +11,7 @@
 
 **What's Fixed**
 * [LazyDataSource]: fixed cascade selection with not flatten search.
+* removed redundant exports
 
 
 # 5.4.2 - 07.12.2023